press3-77 css fixes
