+++ modal.scss
@@ -6,7 +6,7 @@ $gray-text: #394150;
 
     h4,
     h3 {
-      margin: 0 0 !important;
+      margin: 0;
     }
 
     .modal-header {
